Legge til consumer/producer
Legg til Avro consumer som leser fra on prem journalforing topic,
og producer som sender meldingene som JSON til Aiven topic.

navikt/dagpenger#744

+++ src/test/kotlin/no/nav/dagpenger/joark/mottak/DummyJoarkData.kt
@@ -34,7 +34,7 @@ fun lagJoarkHendelse(journalpostId: Long, tema: String, hendelsesType: String, m
     return GenericData.Record(joarkjournalfoeringhendelserAvroSchema).apply {
         put("journalpostId", journalpostId)
         put("hendelsesId", journalpostId.toString())
-        put("versjon", journalpostId)
+        put("versjon", journalpostId.toInt())
         put("hendelsesType", hendelsesType)
         put("journalpostStatus", "journalpostStatus")
         put("temaGammelt", tema)
